Setting up On Cyberspace and Agentics
Adding a Git Repo push for writing and storing Information during the Writing Process of the Teaching Paper:
"On Cyberspace and Agentics" By Robert Watkins Jr.
This tool is being developed to enhance the Citizen Science Data Practices, and Telemetry Data Condensate for future Model Training.
This is part of the Lucid Technologies HITL + AI research database.